BUG FIX: Server errors if posts_dir value contains spaces
If the path specified in posts_dir contained spaces, an exception was thrown in the latest_post_meta method of the MbAnnouncement class in mb_server.py.  This was due to a naive form of pattern matching to identify the latest post.  The code is now based on a more robust regular expression match.