FEATURE: @MB Q to elicit announcements
It's useful to be able to discover MB servers.  You can wait for announcements, but this may mean waiting 60 mins.  This change adds support for a (Q)uery command.  If a server sees a letter Q sent to the @MB group it will immediately generate an @MB announcement.  This process restarts the announcement timer on the server.