Feature: Updated posts, changed default posts_dir and corrected README.md
Anyone wanting to just try the server needed to create a directory for posts, copy sample posts to the directory and update a server setting.  The sample posts started at post 22 and were also out of date.

This version tidies this area and enables the server to work out-of-the-box by defaulting the posts_dir setting to the sample directory.  The sample posts have been updated and start from post id 1.

The README.md file also contained errors.  This has been corrected.