Feature: Support for an upstream web-based blog store
This change adds code that retrieves new posts from an upstream store - in this case a web server.  This is optional; manually entering posts into posts_dir continues to be supported.  The url of the store is defined in the server_settings file.  If posts_url_root contains a string, the upstream store is checked, and if it is NULL there is no action.  Incorrect definition of the URL or an HTTP error (including 404 NOT FOUND) is ignored, although the code does note the error in the console log.